requirements.txt
pandas
kafka-python
requests
tqdm
python-dateutil